MAT 111 Survey

Results of a survey of MAT 111 students at Georgetown College.

height. How tall are you, in inches?

ideal_ht. A numeric vector How tall would you LIKE to be, in inches?

sleep. How much sleep did you get last night?

fastest. What is the highest speed at which you have ever driven a car?

weight_feel. How do you feel about your weight?

love_first. Do you believe in love at first sight?

extra_life. Do you believe in extraterrestrial life?

seat. When you have a choice, where do you prefer to sit in a classroom?

GPA. What is your college GPA?

enough_Sleep. Do you think you get enough sleep?

sex. What sex are you?

diff. Your ideal height minus your actual height.

Format

A data frame with 71 rows and 12 variables

Source

Georgetown College, MAT 111.